# aurora-go/docs/swagger.yaml
basePath: /api
definitions:
  github_com_aurora-go_aurora_internal_dto.AnalyzeRequest:
    properties:
      dateRange:
        example: 7d
        type: string
      metric:
        example: pv,uv,bounce_rate
        type: string
      type:
        example: traffic
        type: string
    required:
    - type
    type: object
...
  github_com_aurora-go_aurora_internal_dto.SearchRequest:
    properties:
      pageNum:
        example: 1
        type: integer
      pageSize:
        example: 10
        type: integer
      query:
        example: Go语言如何实现高并发
        type: string
      topK:
        description: RAG检索数量
        example: 5
        type: integer
      useRag:
        description: 是否启用RAG增强
        example: true
        type: boolean
    required:
    - query
    type: object
  github_com_aurora-go_aurora_internal_dto.WriteRequest:
    properties:
      action:
        example: generate
        type: string
      articleId:
        type: integer
      content:
        example: 现有文章内容...
        type: string
      length:
        example: medium
        type: string
      style:
        example: professional
        type: string
      topic:
        example: Go语言并发编程实战
        type: string
    required:
    - action
    type: object
...
host: localhost:8080
info:
  contact:
    email: aurora@example.com
    name: Aurora Team
  description: Aurora 博客系统 Go 1.26 后端 - 基于 tRPC-Agent-Go 的 AI 驱动博客平台
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/nicepkg/aurora
  title: Aurora Blog API
  version: "1.0"
paths:
...
  /api/about:
    get:
      consumes:
      - application/json
      description: 获取前台公开的关于页面内容
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回关于页面内容
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 获取关于页面内容
      tags:
      - 关于
...
  /api/admin/about:
    post:
      consumes:
      - application/json
      description: 保存或更新关于页面内容（后台管理）
      parameters:
      - description: 关于页面内容
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 保存/更新关于页面
      tags:
      - 关于
...
  /api/agent/analyze:
    post:
      consumes:
      - application/json
      description: 使用AI进行数据分析并获取洞察结果
      parameters:
      - description: 分析请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_dto.AnalyzeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回分析结果
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: 数据分析+AI洞察
      tags:
      - AI Agent
  /api/agent/chat:
    get:
      consumes:
      - application/json
      description: 通过SSE流式输出或同步模式进行AI对话
      parameters:
      - description: 对话消息
        in: query
        name: message
        type: string
      - description: 会话ID
        in: query
        name: sessionId
        type: string
      - description: 对话模式
        in: query
        name: mode
        type: string
      - description: 是否流式输出
        in: query
        name: stream
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回AI响应
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: SSE流式AI对话
      tags:
      - AI Agent
  /api/agent/search:
    post:
      consumes:
      - application/json
      description: 使用AI进行语义搜索
      parameters:
      - description: 搜索请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_dto.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回搜索结果
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: AI语义搜索
      tags:
      - AI Agent
  /api/agent/sessions:
    get:
      consumes:
      - application/json
      description: 获取用户的AI对话会话列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回会话列表
          schema:
            type: object
        "401":
          description: 未授权/Token无效
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: AI对话会话列表
      tags:
      - AI Agent
  /api/agent/write:
    post:
      consumes:
      - application/json
      description: 使用AI生成文章/内容
      parameters:
      - description: 写作请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_aurora-go_aurora_internal_dto.WriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回写作结果
          schema:
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: object
        "500":
          description: 服务器内部错误
          schema:
            type: object
      summary: AI写作助手
      tags:
      - AI Agent
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Token 认证，格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"